LookupServer: Port to LibCore
